@@ -130,38 +130,38 @@
     "@xivapi/nodestone": "^0.2.3",
     "buffer-reader": "^0.1.0",
     "electron-is-dev": "^1.1.0",
-    "electron-log": "^4.4.1",
-    "electron-store": "^5.2.0",
-    "fs-extra": "^7.0.1",
+    "electron-log": "^4.4.5",
+    "electron-store": "^8.0.1",
+    "fs-extra": "^10.0.0",
     "hash-files": "^1.1.1",
-    "is-elevated": "^3.0.0",
-    "jwt-decode": "^2.2.0",
-    "lodash": "^4.17.10",
+    "is-elevated": "^4.0.0",
+    "jwt-decode": "^3.1.2",
+    "lodash": "^4.17.21",
     "ngx-google-analytics": "^11.2.1",
     "request": "^2.88.2",
-    "rxjs": "^6.6.3",
+    "rxjs": "^7.5.2",
     "tslib": "^2.3.1"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^12.1.3",
-    "@angular-devkit/build-angular": "~12.2.10",
-    "@angular-devkit/core": "^12.2.10",
-    "@angular-devkit/schematics": "^12.2.10",
-    "@angular/animations": "^12.2.10",
-    "@angular/cdk": "^12.2.10",
-    "@angular/cli": "^12.2.10",
-    "@angular/common": "^12.2.10",
-    "@angular/compiler": "^12.2.10",
-    "@angular/compiler-cli": "^12.2.10",
-    "@angular/core": "^12.2.10",
+    "@angular-builders/custom-webpack": "^13.0.0",
+    "@angular-devkit/build-angular": "^13.1.4",
+    "@angular-devkit/core": "^13.1.4",
+    "@angular-devkit/schematics": "^13.1.4",
+    "@angular/animations": "13.1.3",
+    "@angular/cdk": "13.1.3",
+    "@angular/cli": "13.1.4",
+    "@angular/common": "13.1.3",
+    "@angular/compiler": "13.1.3",
+    "@angular/compiler-cli": "13.1.3",
+    "@angular/core": "13.1.3",
     "@angular/fire": "^7.2.0",
-    "@angular/flex-layout": "^12.0.0-beta.35",
-    "@angular/forms": "^12.2.10",
-    "@angular/language-service": "^12.2.10",
-    "@angular/platform-browser": "^12.2.10",
-    "@angular/platform-browser-dynamic": "^12.2.10",
-    "@angular/platform-server": "^12.2.12",
-    "@angular/router": "^12.2.10",
+    "@angular/flex-layout": "^13.0.0-beta.36",
+    "@angular/forms": "13.1.3",
+    "@angular/language-service": "13.1.3",
+    "@angular/platform-browser": "13.1.3",
+    "@angular/platform-browser-dynamic": "13.1.3",
+    "@angular/platform-server": "13.1.3",
+    "@angular/router": "13.1.3",
     "@ffxiv-teamcraft/simulator": "^3.0.11",
     "@ffxiv-teamcraft/simulator-cn": "npm:@ffxiv-teamcraft/simulator@2.10.12",
     "@ffxiv-teamcraft/simulator-kr": "npm:@ffxiv-teamcraft/simulator@2.10.12",
@@ -256,39 +255,33 @@
     "karma-coverage-istanbul-reporter": "~1.4.2",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "less": "^3.11.1",
+    "less": "^4.1.2",
     "memoizee": "^0.4.14",
     "multi-progress": "^2.0.0",
     "ng-drag-drop": "^5.0.0",
     "ng-push-ivy": "^1.0.8",
-    "ng-zorro-antd": "11.1.0",
+    "ng-zorro-antd": "13.0.1",
     "ngx-color-picker": "^11.0.0",
     "ngx-markdown": "^12.0.1",
     "nz-tslint-rules": "^0.901.2",
-    "ora": "^5.2.0",
     "p-queue": "^7.1.0",
     "papaparse": "^5.2.0",
-    "prettier": "^2.0.5",
+    "prettier": "2.3.1",
     "quicktype-core": "^6.0.70",
-    "redis": "^4.0.1",
     "semver": "^5.4.1",
     "standard-version": "^4.4.0",
     "subscriptions-transport-ws": "^0.9.19",
     "terser-webpack-plugin": "^1.2.3",
     "ts-loader": "^9.2.6",
-    "ts-node": "~9.0.0",
-    "tslint": "~6.1.3",
-    "typescript": "~4.3.5",
+    "ts-node": "9.1.1",
+    "tslint": "6.1.3",
+    "typescript": "4.5.5",
     "utf-8-validate": "^5.0.7",
     "validate-commit-msg": "^2.14.0",
-    "web-animations-js": "^2.3.2",
-    "webpack": "^5.60.0",
-    "webpack-bundle-analyzer": "^4.5.0",
-    "webpack-cli": "^4.9.1",
     "ws": "^8.2.3",
     "xhr2": "^0.2.1",
     "yargs": "^17.1.1",
-    "zone.js": "~0.11.4"
+    "zone.js": "0.11.4"
   },
   "resolutions": {
     "@types/node": "14.14.45"
