@@ -20,6 +20,7 @@ export interface AuthState {
   linkingCharacter: boolean;
   commissionProfile: CommissionProfile;
   logTracking: LogTracking;
+  serverLogTracking?: LogTracking;
 }
 
 export const initialState: AuthState = {
@@ -266,7 +267,8 @@ export function authReducer(state = initialState, action: AuthActions): AuthStat
     case AuthActionTypes.LogTrackingLoaded:
       return {
         ...state,
-        logTracking: action.payload
+        logTracking: action.payload,
+        serverLogTracking: action.payload
       };
 
     case AuthActionTypes.MarkAsDoneInLog:
