@@ -14,7 +14,6 @@ import { I18nToolsService } from '../../../../core/tools/i18n-tools.service';
 import {
   debounceTime,
   distinctUntilChanged,
-  exhaustMap,
   filter,
   first,
   map,
@@ -126,10 +125,6 @@ export class ItemRowComponent extends TeamcraftComponent implements OnInit {
   team$: Observable<Team>;
 
   missingBooks$: Observable<number[]>;
-
-  commentBadge$: Observable<boolean>;
-
-  commentBadgeReloader$: BehaviorSubject<void> = new BehaviorSubject<void>(null);
 
   tagInputVisible = false;
 
@@ -372,19 +367,6 @@ export class ItemRowComponent extends TeamcraftComponent implements OnInit {
     ).subscribe(value => {
       this.itemDoneChanged(value, this.item);
     });
-
-    this.commentBadge$ = this.commentBadgeReloader$.pipe(
-      exhaustMap(() => combineLatest([this.list$, this.item$.pipe(map(i => i.id))])),
-      switchMap(([list, itemId]) => {
-        return this.commentsService.getComments(
-          CommentTargetType.LIST,
-          list.$key,
-          `${this.finalItem ? 'finalItems' : 'items'}:${itemId}`
-        );
-      }),
-      map(comments => comments.length > 0),
-      startWith(false)
-    );
   }
 
   openMarketboardDialog(item: ListRow): void {
@@ -578,9 +560,7 @@ export class ItemRowComponent extends TeamcraftComponent implements OnInit {
           })
         );
       })
-    ).subscribe(() => {
-      this.commentBadgeReloader$.next(null);
-    });
+    ).subscribe();
   }
 
   setWorkingOnIt(uid: string, item: ListRow): void {
